[EP-2124] Search and sidebar refactoring
