+++ packages/earth-map/src/components/layers/layer/styles.scss
@@ -1,6 +1,6 @@
 @import '~styles/config';
 
-.c-sidebar-panel {
+.c-sidebar {
 
   &.no-scroll {
     overflow: hidden;
