+++ packages/earth-map/src/components/org-switcher/component.tsx
@@ -1,11 +1,14 @@
 import React, { useContext, useState } from 'react';
-import { Auth0Context } from 'auth/auth0';
-
-import { checkRole } from 'utils';
-import DropdownComponent from 'components/dropdown';
-import './styles.scss';
 import { remove } from 'lodash';
 import classNames from 'classnames';
+import Link from 'redux-first-router-link';
+
+import { Auth0Context } from 'auth/auth0';
+import { checkRole } from 'utils';
+import DropdownComponent from 'components/dropdown';
+import { APP_NAME, APP_LOGO } from '../../theme';
+
+import './styles.scss';
 
 const ADMIN_PATH = process.env.REACT_APP_ADMIN_URL;
 
@@ -14,7 +17,7 @@ const OrgSwitcher = (props) => {
     roles,
     userData: { allGroups },
   } = useContext(Auth0Context);
-  const { resetPlacesFeatured, resetLayerCache } = props;
+  const { resetPlacesFeatured, resetLayerCache, resetMap } = props;
   const { group, setUserGroup, setPlacesSearch } = props;
   const hasMultipleGroups = allGroups.length > 1;
   const allInitiallySelected = group.length === allGroups.length;
@@ -53,69 +56,90 @@ const OrgSwitcher = (props) => {
   };
 
   return (
-    <>
-      <span className="ng-text-display-s ng-text-weight-regular ng-body-color ng-margin-remove ng-display-block ng-org-name">
-        map view
+
+    <div className="ng-padding-medium-horizontal ng-ep-background-dark ng-flex ng-flex-middle ng-position-relative ng-padding-bottom ng-padding-small-top">
+      <Link
+        className="ng-border-remove"
+        to={{
+          type: 'EARTH',
+        }}
+      >
+        <img
+          src={APP_LOGO}
+          alt={APP_NAME}
+          className="ng-margin-remove ng-display-block"
+          onClick={resetMap}
+        />
+      </Link>
+      <span className="ng-ep-kicker"></span>
+
+      {/* Old org switcher */}
+      <>
+        <span className="ng-text-display-s ng-text-weight-regular ng-body-color ng-margin-remove ng-display-block ng-org-name">
+          map view
       </span>
 
-      <div
-        onClick={handleDropdownToggle}
-        className="ng-padding ng-c-cursor-pointer ng-position-relative"
-      >
-        <i
-          className={classNames({
-            'ng-body-color': true,
-            'ng-icon-directionup': dropdownState === 'open',
-            'ng-icon-directiondown': dropdownState !== 'open',
-          })}
-        />
-        {selectedGroups.length > 0 && <span className="ng-org-badge">{selectedGroups.length}</span>}
-      </div>
+        <div
+          onClick={handleDropdownToggle}
+          className="ng-padding ng-c-cursor-pointer ng-position-relative"
+        >
+          <i
+            className={classNames({
+              'ng-body-color': true,
+              'ng-icon-directionup': dropdownState === 'open',
+              'ng-icon-directiondown': dropdownState !== 'open',
+            })}
+          />
+          {selectedGroups.length > 0 && <span className="ng-org-badge">{selectedGroups.length}</span>}
+        </div>
 
-      <DropdownComponent state={dropdownState}>
-        <li className="ng-ep-dropdown-category ng-ep-dropdown-selected">
-          <span className="ng-dropdown-item">MAP VIEW</span>
-        </li>
-        <li className="ng-form ng-form-dark">
-          <div className="ng-padding-medium-horizontal ng-padding-top">
-            {Object.keys(roles).map((g, i) => (
-              <label
-                htmlFor={g}
-                className={classNames({
-                  'ng-display-block ng-padding-bottom': true,
-                  'ng-c-cursor-pointer': hasMultipleGroups,
-                })}
-                key={i}
-              >
-                {hasMultipleGroups && (
-                  <input
-                    className="ng-checkbox-input"
-                    type="checkbox"
-                    id={g}
-                    value={g}
-                    checked={!!selectedGroups.find((x) => x === g)}
-                    name={g}
-                    onChange={(e) => onOrgChange(e)}
-                  />
-                )}
-                {g}
-              </label>
-            ))}
-          </div>
-        </li>
+        <DropdownComponent state={dropdownState}>
+          <li className="ng-ep-dropdown-category ng-ep-dropdown-selected">
+            <span className="ng-dropdown-item">MAP VIEW</span>
+          </li>
+          <li className="ng-form ng-form-dark">
+            <div className="ng-padding-medium-horizontal ng-padding-top">
+              {Object.keys(roles).map((g, i) => (
+                <label
+                  htmlFor={g}
+                  className={classNames({
+                    'ng-display-block ng-padding-bottom': true,
+                    'ng-c-cursor-pointer': hasMultipleGroups,
+                  })}
+                  key={i}
+                >
+                  {hasMultipleGroups && (
+                    <input
+                      className="ng-checkbox-input"
+                      type="checkbox"
+                      id={g}
+                      value={g}
+                      checked={!!selectedGroups.find((x) => x === g)}
+                      name={g}
+                      onChange={(e) => onOrgChange(e)}
+                    />
+                  )}
+                  {g}
+                </label>
+              ))}
+            </div>
+          </li>
 
-        {Object.keys(roles).map(
-          (g, i) =>
-            checkRole(roles[g]) && (
-              <li className="ng-ep-dropdown-category" key={i}>
-                <a href={`${ADMIN_PATH}${g}`} className="ng-c-cursor-pointer ng-dropdown-item">
-                  {g} - ADMIN
+          {Object.keys(roles).map(
+            (g, i) =>
+              checkRole(roles[g]) && (
+                <li className="ng-ep-dropdown-category" key={i}>
+                  <a href={`${ADMIN_PATH}${g}`} className="ng-c-cursor-pointer ng-dropdown-item">
+                    {g} - ADMIN
                 </a>
-              </li>
-            )
-        )}
-      </DropdownComponent>
-    </>
+                </li>
+              )
+          )}
+        </DropdownComponent>
+      </>
+      {/* Old org switcher */}
+
+    </div>
   );
 };
