--- packages/earth-map/src/components/sidebar-panel/styles.scss
@@ -1,46 +1,0 @@
-@import '~styles/config';
-
-.c-sidebar-panel {
-  box-shadow: 0 0 5px 0 $marapp-gray-9;
-  position: absolute;
-  display: block;
-  width: 375px;
-  height: 100%;
-  top: 0;
-  left: $c-sidebar-width;
-  z-index: $z-index-2;
-  opacity: 0;
-  overflow: auto;
-
-  .sidebar-panel--btn-close {
-    position: absolute;
-    top: $space-1 * 2.5;
-    right: $space-1 * 3;
-    fill: $marapp-gray-100;
-    z-index: $z-index-3;
-
-    > svg {
-      display: block;
-    }
-  }
-  &.expanded {
-    width: 500px;
-  }
-}
-
-.ng-ep-kicker {
-  margin-left: 10px;
-  margin-right: 15px;
-  position: relative;
-  &:before {
-    content: '';
-    background: $marapp-gray-0;
-    display: block;
-    height: 10px;
-    position: absolute;
-    transform: translateY(-50%);
-    top: 0;
-    bottom: 0;
-    width: 2px;
-  }
-}
