+++ packages/earth-map/src/components/sidebar/sidebar-close/index.tsx
@@ -1,0 +1,2 @@
+export { default } from './component';
+
