+++ packages/earth-map/src/components/sidebar/sidebar-layout-search/index.ts
@@ -1,0 +1,9 @@
+import { connect } from 'react-redux';
+import SidebarLayoutSearch from './component';
+
+export default connect(
+  (state: any) => ({
+    ...state.indexes,
+    ...state.sidebar,
+  })
+)(SidebarLayoutSearch);
