+++ packages/earth-map/src/components/sidebar/styles.scss
@@ -1,18 +1,44 @@
 @import '~styles/config';
 
 .c-sidebar {
+  box-shadow: 0 0 5px 0 $marapp-gray-9;
   position: absolute;
-  display: flex;
-  width: 80px;
-  top: 59px;
-  left: 0;
+  display: block;
+  width: 375px;
+  top: 0;
+  bottom: 0;
+  left: $c-sidebar-width;
   z-index: $z-index-3;
 
-  .sidebar--btn {
-    height: 50px;
-    width: 40px;
-    background-color: $marapp-gray-8;
-    box-shadow: 0 0 5px 0 rgba($marapp-gray-100, 0.5);
-    transition: all 0.3s ease;
+  .sidebar-panel--btn-close {
+    position: absolute;
+    top: $space-1 * 2.5;
+    right: $space-1 * 3;
+    fill: $marapp-gray-100;
+    z-index: $z-index-3;
+
+    > svg {
+      display: block;
+    }
+  }
+  &.expanded {
+    width: 500px;
+  }
+}
+
+.ng-ep-kicker {
+  margin-left: 10px;
+  margin-right: 15px;
+  position: relative;
+  &:before {
+    content: '';
+    background: $marapp-gray-0;
+    display: block;
+    height: 10px;
+    position: absolute;
+    transform: translateY(-50%);
+    top: 0;
+    bottom: 0;
+    width: 2px;
   }
 }
